Update wrapper to take multi species output config
The wrapper now works with multispecies runs with just sapflux. There isn't a function to calculate the metric for plot transpiration yet, so that is commented out in the config.

The wrapper takes the species and plot parameters and converts them to the format that boa needs, and then when it gets them back from boa, converts them back. It also passes a series of properties from the objectives with each objective to the wrapper for calculating the sapflux RMSE.